perf(ci): optimize performance tests execution duration
- Refactor NBomber test classes to use dynamic configuration durations

- Add Category xUnit Traits directly to individual performance test classes

- Remove recursive test execution loop from PerformanceTestSuite

- Set environment variables to scale test durations to seconds for PR triggers

+++ backend/tests/Footex.PerformanceTests/Common/PerformanceTestSettings.cs
@@ -43,22 +43,22 @@ public class DurationSettings
     /// <summary>
     ///     Short test duration in minutes
     /// </summary>
-    public int ShortTestMinutes { get; set; } = 1;
+    public double ShortTestMinutes { get; set; } = 1;
 
     /// <summary>
     ///     Medium test duration in minutes
     /// </summary>
-    public int MediumTestMinutes { get; set; } = 2;
+    public double MediumTestMinutes { get; set; } = 2;
 
     /// <summary>
     ///     Long test duration in minutes
     /// </summary>
-    public int LongTestMinutes { get; set; } = 5;
+    public double LongTestMinutes { get; set; } = 5;
 
     /// <summary>
     ///     Stress test duration in minutes
     /// </summary>
-    public int StressTestMinutes { get; set; } = 10;
+    public double StressTestMinutes { get; set; } = 10;
 }
 
 public class LoadSettings
@@ -104,7 +104,7 @@ public class CacheSettings
     /// <summary>
     ///     Cache hit test duration in minutes
     /// </summary>
-    public int CacheHitTestMinutes { get; set; } = 2;
+    public double CacheHitTestMinutes { get; set; } = 2;
 
     /// <summary>
     ///     Expected cache hit ratio (0.0 to 1.0)
